Change pedestals fluid-xp conversions.
Pedestals config changes:

* Limit break the xp anvil. This is a skyblock pack, Smite XX just isn't that big a deal and folks will enjoy the challenge.
* Remove the default fluid->xp conversion recipes.
* Add fluid xp conversion for menril, fluid xps, and blood magic essence
* Significantly nerf Lava's XP value since it's so trivially available in this pack.

The Lava change, as I test it, makes it so that it's possible to save up enough xp for some low level enchants, but effectively impossible to use lava for significant XP gain.

This change has the implication of turning Blood Magic wells of suffering and Sacrifice setups into substantial sources of EXP for enchanting. It also incentivizes a large menril squeezing farm for enchantments as well.